+++ lib/pats/categories/psg_overall_study_qualities/psg_overall_study_quality_3.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require 'pats/categories/default'
+
+module Pats
+  module Categories
+    module PSGOverallStudyQualities
+      # Defines PSG Overall Study Quality
+      class PSGOverallStudyQuality3 < Default
+        def label
+          '3: Very Good'
+        end
+
+        def variable_name
+          'psg_overall_study_quality'
+        end
+
+        def subquery
+          "NULLIF(response, '')::numeric = 3"
+        end
+      end
+    end
+  end
+end