+++ lib/pats/categories/psg_study_passed/all.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'pats/categories/psg_study_passed/psg_study_passed_yes'
+require 'pats/categories/psg_study_passed/psg_study_passed_no'
+require 'pats/categories/psg_study_passed/psg_study_passed_unknown'